Refactor: move components and utils for cluster creation based on app bundles to own folder
