+++ src/components/MAPI/clusters/CreateClusterAppBundles/custom-widgets/InstanceTypeWidget.tsx
@@ -4,7 +4,7 @@ import { Box, Text } from 'grommet';
 import React from 'react';
 import BaseInputTemplate from 'UI/JSONSchemaForm/BaseInputTemplate';
 
-import { ICreateClusterFormContext } from '../CreateClusterAppBundles';
+import { ICreateClusterFormContext } from '..';
 
 const EMPTY_VALUE_NAME = 'None';
